app/requirements.txt
fastapi
uvicorn
kubernetes
requests
google-generativeai
